Reparaciónes y atomicación del codigo, Juntar botones del login y registro en un nuevo componente  ✔
Agregar el boton de notificaciones en el header ✔

Modificaciónla pagina 404 con el nuevo sistema de css :root ✔

Atomificación del header funcional ✔

Modificación del css del header al nuevo sistema con :root ✔
