+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import AuthPage from './components/AuthPage';
 import NotFound from './components/AppUser/Layout/NotFound';
 import Layout from './components/AppUser/Layout';
-import ProfileEditForm from './components/AppUser/Layout/Header/Display/Profile';
+import ProfileEditForm from './components/AppUser/Layout/Header/UserIcon/UserModal/Profile';
 import HomePage from './components/AppUser/Body/HomePage';
 import Deck from './components/AppUser/Body/DeckPage';
 import User from './components/AppUser/Body/UserPage';
