--- frontend/src/components/AppUser/Layout/Header/Display/Profile/profile.module.css
@@ -1,63 +1,0 @@
-.formContainer {
-    max-width: 400px;
-    margin: 20px auto;
-    padding: 20px;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    background-color: #f9f9f9;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.formTitle {
-    font-size: 1.5rem;
-    margin-bottom: 20px;
-    color: #333;
-    text-align: center;
-}
-
-.formGroup {
-    margin-bottom: 15px;
-    display: flex;
-    flex-direction: column;
-}
-
-.label {
-    font-weight: bold;
-    margin-bottom: 5px;
-    color: #555;
-}
-
-.input {
-    padding: 8px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    font-size: 1rem;
-    transition: border-color 0.2s;
-}
-
-.input:focus {
-    border-color: #007bff;
-    outline: none;
-}
-
-.error {
-    color: #ff4d4f;
-    margin-top: 5px;
-    font-size: 0.875rem;
-}
-
-.button {
-    padding: 10px 15px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 1rem;
-    width: 100%;
-    transition: background-color 0.2s;
-}
-
-.button:hover {
-    background-color: #0056b3;
-}
