+++ frontend/src/components/AppUser/Layout/Header/HeaderIcon/headerIcon.module.css
@@ -1,0 +1,10 @@
+.gameIcon {
+  width: 5rem;
+  height: auto;
+  cursor: pointer;
+  transition: transform 0.4s ease-in-out;
+}
+
+.gameIcon:hover {
+  transform: scale(1.1);
+}
