+++ frontend/src/components/AppUser/Layout/Header/HeaderNav/headerNav.module.css
@@ -1,0 +1,24 @@
+.nav {
+  display: flex;
+  gap: 1.8rem;
+  margin-top: 12px;
+}
+
+.navLink {
+  padding: 0 8px;
+  border-bottom: 1px solid transparent;
+  text-decoration: none;
+  color: var(--color-text-light);
+  font-size: 16px;
+  transition:
+    color 0.3s ease,
+    transform 0.2s ease-in-out;
+}
+
+.navLink:hover {
+  padding: 0 8px;
+  color: var(--color-text-light);
+  border-bottom: 1px solid var(--color-hover);
+  border-radius: 0 0 8px 8px;
+  transform: translateY(-6px);
+}
