+++ frontend/src/components/AppUser/Layout/Header/UserIcon/UserModal/userModal.module.css
@@ -1,0 +1,33 @@
+.modalUser {
+  display: flex;
+  flex-direction: column;
+  position: absolute;
+  top: 80px;
+  right: 0;
+  padding: 5px;
+  width: 220px;
+  background: var(--color-background-header);
+  border: 1px solid var(--color-border-1);
+  border-radius: 8px;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.3);
+  z-index: 10;
+}
+
+.modalUser button {
+  padding: 10px;
+  width: 100%;
+  background: none;
+  border: none;
+  border-radius: 50px;
+  text-align: left;
+  font-size: 14px;
+  font-family: var(--font-family);
+  color: var(--color-text-light);
+  cursor: pointer;
+  transition: background 0.2s ease-in-out;
+}
+
+.modalUser button:hover {
+  background: var(--color-hover);
+  border-radius: 50px;
+}
