+++ frontend/src/components/AppUser/Layout/Header/UserIcon/userIcon.module.css
@@ -1,0 +1,25 @@
+.userContainer {
+  position: relative;
+  display: flex;
+  gap: 10px;
+  align-items: center;
+  cursor: pointer;
+}
+
+.notification {
+  margin-top: 20px;
+  font-size: 1.5rem;
+  color: var(-color-text-light);
+}
+
+.userIconImage img {
+  width: 60px;
+  height: 60px;
+  border-radius: 50%;
+  object-fit: cover;
+  transition: transform 0.4s ease-in-out;
+}
+
+.userIconImage:hover img {
+  transform: scale(1.1);
+}
