+++ frontend/src/components/AppUser/Layout/index.jsx
@@ -1,7 +1,6 @@
 import '@fontsource/metamorphous';
 import Header from './Header';
 import { Outlet } from 'react-router-dom';
-import Footer from './Footer';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
@@ -12,7 +11,6 @@ const Layout = () => (
     <main>
       <Outlet />
     </main>
-    {/* <Footer /> */}
   </>
 );
