+++ frontend/src/components/AuthPage/AuthFormSelector/authFormSelector.module.css
@@ -8,7 +8,7 @@
   min-width: 80%;
   padding: 0.5rem;
 
-  background-color: #f8edd3;
+  background-color: var(--color-button-bg);
   border-radius: 20px;
 }
 
@@ -22,12 +22,12 @@
   color: #77767e;
   cursor: pointer;
   transition: background-color 0.3s;
-  font-family: 'Metamorphous', sans-serif;
+  font-family: var(--font-family);
 }
 
 .tab.active {
   border-radius: 20px 0 0 20px;
-  border-bottom: 2px solid #c6a65f;
+  border-bottom: 2px solid var(--color-active);
 
   font-size: 0.7rem;
   color: rgb(0, 0, 0);
@@ -36,7 +36,7 @@
 
 .tab.active2 {
   border-radius: 0 20px 20px 0;
-  border-bottom: 2px solid #c6a65f;
+  border-bottom: 2px solid var(--color-active);
 
   font-size: 0.7rem;
   color: rgb(0, 0, 0);
