+++ frontend/src/components/AuthPage/AuthTitle/authtitle.module.css
@@ -1,12 +1,12 @@
 .header {
-  color: rgb(255, 255, 255);
+  color: var(--color-light);
   text-align: center;
   padding: 1rem 0;
-  font-family: 'Metamorphous', sans-serif;
+  font-family: var(--font-family);
 }
 
 .bg-6 {
-  background-color: #664631;
+  background-color: var(--color-dark);
   display: flex;
   justify-content: center;
 }
@@ -15,7 +15,7 @@
   font-size: 1.6rem;
   position: relative;
   margin: 0 auto;
-  font-family: 'Pixelify Sans', sans-serif;
+  font-family: var(--font-family-2);
 }
 
 .glitch:after {
@@ -23,7 +23,7 @@
   position: absolute;
   left: 12vw;
   text-shadow: -2px 0 #8a2d16;
-  background: #664631;
+  background: var(--color-dark);
   overflow: hidden;
   clip: rect(0, 900px, 0, 0);
   animation: noise-anim 3s infinite linear alternate-reverse;
@@ -33,8 +33,8 @@
   content: attr(data-text);
   position: absolute;
   left: 16vw;
-  text-shadow: 1px 0 #53612b;
-  background: #664631;
+  text-shadow: 1px 0 var(--color-primary);
+  background: var(--color-dark);
   overflow: hidden;
   clip: rect(0, 900px, 0, 0);
   animation: noise-anim-2 3s infinite linear alternate-reverse;
