+++ frontend/src/components/AuthPage/LoginForm/loginform.module.css
@@ -16,11 +16,11 @@
 .field input {
   padding: 8px;
 
-  border: 1px solid #f8edd3;
+  border: 1px solid var(--color-button-bg);
   border-radius: 6px;
-  background-color: #f8edd3;
+  background-color: var(--color-button-bg);
 
-  font-family: 'Metamorphous', sans-serif;
+  font-family: var(--font-family);
   font-size: 10px;
   transition: border-color 0.3s ease-in-out;
 }
@@ -30,34 +30,6 @@
   font-size: 0.6rem;
   margin-top: 0.4rem;
   margin-left: 0.4rem;
-}
-
-.loginButton {
-  flex: 1;
-  margin-top: 20px;
-  padding: 8px;
-  font-size: 12px;
-  font-weight: bold;
-  font-family: 'Metamorphous', sans-serif;
-
-  border: none;
-  border-bottom: 2px solid #f8edd3;
-  border-radius: 0 0 20px 20px;
-
-  background-color: transparent;
-  cursor: pointer;
-  color: #9fb033;
-  transition:
-    transform 0.2s ease,
-    background-color 0.3s ease-in-out;
-}
-
-.loginButton:hover {
-  transform: translateY(-4px);
-}
-
-.loginButton:active {
-  transform: translateY(0);
 }
 
 @media screen and (min-width: 576px) {
@@ -80,46 +52,3 @@
     font-size: 14px;
   }
 }
-
-@media screen and (min-width: 992px) {
-}
-
-@media screen and (min-width: 1200px) {
-}
-
-@media screen and (min-width: 1800px) {
-  .field input {
-    font-size: 14px;
-  }
-
-  .loginButton {
-    font-size: 16px;
-  }
-}
-
-/* 
-
-
-
-
-.field input:focus {
-  border-color: #838383;
-  outline: none;
-}
-
-
-
-.actions {
-  display: flex;
-  justify-content: center;
-}
-
-@media (max-width: 575.98px) {
-  .form {
-    padding: 1rem;
-  }
-
-  .field input {
-    font-size: 0.9rem;
-  }
-} */
