+++ frontend/src/components/AuthPage/RegisterForm/registerform.module.css
@@ -18,11 +18,11 @@
 .field input {
   padding: 8px;
 
-  border: 1px solid #f8edd3;
+  border: 1px solid var(--color-button-bg);
   border-radius: 6px;
-  background-color: #f8edd3;
+  background-color: var(--color-button-bg);
 
-  font-family: 'Metamorphous', sans-serif;
+  font-family: var(--font-family);
   font-size: 10px;
   transition: border-color 0.3s ease-in-out;
 }
@@ -45,48 +45,17 @@
 .field span {
   flex: 0.75;
   font-size: 10px;
-  font-family: 'Metamorphous', sans-serif;
+  font-family: var(--font-family);
 }
 
 .error {
-  color: #d9534f;
+  color: var(--color-error);
   font-size: 0.6rem;
   margin-top: 0.4rem;
   margin-left: 0.4rem;
 }
 
-.registerButton {
-  flex: 1;
-  padding: 6px;
-  font-weight: bold;
-
-  border: none;
-  border-bottom: 2px solid #f8edd3;
-  border-radius: 0 0 20px 20px;
-  background-color: transparent;
-
-  color: #9fb033;
-  font-size: 12px;
-  font-family: 'Metamorphous', sans-serif;
-  cursor: pointer;
-  transition:
-    transform 0.2s ease,
-    background-color 0.3s ease-in-out;
-}
-
-.registerButton:hover {
-  transform: translateY(-4px);
-}
-
-.registerButton:active {
-  transform: translateY(0);
-}
-
 @media screen and (min-width: 576px) {
-  .registerButton {
-    margin-top: 12px;
-    font-size: 14px;
-  }
 }
 
 @media screen and (min-width: 768px) {
@@ -100,10 +69,6 @@
 
   .field span {
     font-size: 11px;
-  }
-
-  .registerButton {
-    font-size: 14px;
   }
 }
 
@@ -120,9 +85,5 @@
 
   .field span {
     font-size: 12px;
-  }
-
-  .registerButton {
-    font-size: 16px;
   }
 }
