+++ frontend/src/components/AuthPage/index.jsx
@@ -1,9 +1,11 @@
 import styles from './authpage.module.css';
-import { useState } from 'react';
 import AuthTitle from './AuthTitle';
 import AuthFormSelector from './AuthFormSelector';
 import LoginForm from './LoginForm';
 import RegisterForm from './RegisterForm';
+import { useState } from 'react';
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 
 const AuthPage = ({ forceUpdate, setForceUpdate }) => {
   const [showLogin, setShowLogin] = useState(true);
@@ -11,19 +13,14 @@ const AuthPage = ({ forceUpdate, setForceUpdate }) => {
   return (
     <div className={styles.container}>
       <div className={styles.authContent}>
+        <ToastContainer />
         <AuthTitle />
         <AuthFormSelector showLogin={showLogin} setShowLogin={setShowLogin} />
         <div className={styles.formContainer}>
           {showLogin ? (
-            <LoginForm
-              forceUpdate={() => setForceUpdate(!forceUpdate)}
-              onSubmit={(data) => loginMutation.mutate(data)}
-            />
+            <LoginForm forceUpdate={() => setForceUpdate(!forceUpdate)} />
           ) : (
-            <RegisterForm
-              forceUpdate={() => setForceUpdate(!forceUpdate)}
-              onSubmit={(data) => registerMutation.mutate(data)}
-            />
+            <RegisterForm forceUpdate={() => setForceUpdate(!forceUpdate)} />
           )}
         </div>
       </div>
