+++ frontend/src/components/AuthPage/sendButton/sendButton.module.css
@@ -1,0 +1,27 @@
+.authButton {
+  flex: 1;
+  margin-top: 20px;
+  padding: 8px;
+  font-size: 12px;
+  font-weight: bold;
+  font-family: var(--font-family);
+
+  border: none;
+  border-bottom: 2px solid var(--color-button-bg);
+  border-radius: 0 0 20px 20px;
+
+  background-color: transparent;
+  cursor: pointer;
+  color: var(--color-button-send);
+  transition:
+    transform 0.2s ease,
+    background-color 0.3s ease-in-out;
+}
+
+.authButton:hover {
+  transform: translateY(-4px);
+}
+
+.authButton:active {
+  transform: translateY(0);
+}
