+++ frontend/src/index.css
@@ -26,3 +26,20 @@ h4 {
 p {
   text-wrap: pretty;
 }
+
+:root {
+  --color-background-header: #521e00;
+  --color-primary: #54612b;
+  --color-primary-hover: #373f1c;
+  --color-light: #ffffff;
+  --color-dark: #664631;
+  --color-text-light: #f5f5f5;
+  --color-border-1: #b87937;
+  --color-button-bg: #f8edd3;
+  --color-button-send: #9fb033;
+  --color-hover: #a87304;
+  --color-active: #c6a65f;
+  --color-error: #d9534f;
+  --font-family: 'Metamorphous', sans-serif;
+  --font-family-2: 'Pixelify Sans', sans-serif;
+}
