+++ frontend/src/lib/utils/apiCard.js
@@ -6,16 +6,8 @@ const API = axios.create({
 });
 
 export const fetchCards = async () => {
-  try {
-    const response = await API.get('/');
-    const data = response.data;
-    if (!Array.isArray(data)) {
-      return [];
-    }
-    return data;
-  } catch (error) {
-    return [];
-  }
+  const response = await API.get('/');
+  return response.data;
 };
 
 export const createCard = async (card) => {
