+++ frontend/src/lib/utils/apiUser.js
@@ -80,5 +80,5 @@ export const loginUser = async (data) => {
 
 export const registerUser = async (data) => {
   const response = await API.post('/auth/register', data);
-  return response.data;
+  return response;
 };
