+++ frontend/src/lib/utils/apiUserCollection.js
@@ -7,18 +7,16 @@ const API = axios.create({
 
 export const fetchUserCollection = async () => {
   const token = getUserToken();
-  try {
-    const response = await API.get('/', {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data.cards || [];
-  } catch (e) {
-    return [];
-  }
+
+  const response = await API.get('/', {
+    headers: { Authorization: `Bearer ${token}` },
+  });
+  return response.data.cards;
 };
 
 export const createCardUser = (idUser, idCard) => {
   const token = getUserToken();
+
   return API.post(
     '/',
     { userId: idUser, cardId: idCard },
@@ -32,6 +30,7 @@ export const createCardUser = (idUser, idCard) => {
 
 export const deleteCardUser = (idUser, idCard) => {
   const token = getUserToken();
+
   return API.delete(`/${idUser}/cards/${idCard}`, {
     headers: {
       Authorization: `Bearer ${token}`,
